requirements.txt
gradio==5.31.0
Pillow==10.4.0
python-dotenv==1.1.1
zhipuai==2.1.5.20250708
